split bundle into separate files
