+++ .gitignore
@@ -3,3 +3,4 @@ node_modules/
 build
 coverage
 cypress/videos
+dist
