+++ cypress.json
@@ -1,3 +1,3 @@
 {
-  "baseUrl": "http://localhost:8080"
+  "baseUrl": "http://localhost:1234"
 }
