+++ package.json
@@ -6,8 +6,8 @@
   "scripts": {
     "test": "cypress run",
     "cy:open": "cypress open",
-    "start": "http-server -c-1 -p 8080",
-    "dev": "start-test 8080 cy:open"
+    "start": "parcel serve index.html",
+    "dev": "start-test 1234 cy:open"
   },
   "repository": {
     "type": "git",
@@ -23,6 +23,7 @@
   "devDependencies": {
     "cypress": "3.2.0",
     "http-server": "0.11.1",
+    "parcel-bundler": "1.12.3",
     "prettier": "1.17.0",
     "start-server-and-test": "1.9.0"
   }
